wallet: Factor out test node setup
