+++ Cargo.toml
@@ -63,6 +63,7 @@ members = [
   "wallet/wallet-node-client",          # Wallet-to-node communication tools.
   "wallet/wallet-address-generator",    # Wallet address generator binary.
   "wallet/wallet-address-generator-lib",# Wallet address generator lib.
+  "wallet/wallet-test-node",            # Node for wallet testing as a library.
   "wasm-crypto",                        # WASM bindings for the crypto crate.
 ]
